use similar names for similar functions
